revenue-leak-detection/requirements.txt
pandas==2.2.2
pyarrow==16.1.0
numpy==1.26.4

mysql-connector-python==8.4.0
SQLAlchemy==2.0.30
PyMySQL==1.1.1

dagster==1.7.13
dagster-webserver==1.7.13
dagster-pandas==0.23.13

requests==2.32.3
tqdm==4.66.4

python-dotenv==1.0.1

ruff==0.4.10
